Moving file related stuff into feature folder
