+++ app/src/lib/components/BranchLane.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import BranchCard from './BranchCard.svelte';
-	import FileCard from './FileCard.svelte';
 	import { Project } from '$lib/backend/projects';
 	import Resizer from '$lib/components/Resizer.svelte';
 	import { projectLaneCollapsed } from '$lib/config/config';
+	import FileCard from '$lib/file/FileCard.svelte';
 	import { persisted } from '$lib/persisted/persisted';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
 	import { getContext, getContextStoreBySymbol, createContextStore } from '$lib/utils/context';
