+++ app/src/lib/file/FileContextMenu.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-	import Button from './Button.svelte';
-	import Modal from './Modal.svelte';
 	import { Project } from '$lib/backend/projects';
+	import Button from '$lib/components/Button.svelte';
+	import Modal from '$lib/components/Modal.svelte';
 	import PopupMenu from '$lib/components/PopupMenu.svelte';
 	import ContextMenu from '$lib/components/contextmenu/ContextMenu.svelte';
 	import ContextMenuItem from '$lib/components/contextmenu/ContextMenuItem.svelte';
