+++ app/src/lib/file/FileStatusCircle.svelte
