+++ app/src/lib/file/FileStatusIcons.svelte
