+++ app/src/routes/[projectId]/base/+page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import { Project } from '$lib/backend/projects';
 	import BaseBranch from '$lib/components/BaseBranch.svelte';
-	import FileCard from '$lib/components/FileCard.svelte';
 	import FullviewLoading from '$lib/components/FullviewLoading.svelte';
 	import Resizer from '$lib/components/Resizer.svelte';
 	import ScrollableContainer from '$lib/components/ScrollableContainer.svelte';
+	import FileCard from '$lib/file/FileCard.svelte';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
 	import { getContext, getContextStoreBySymbol } from '$lib/utils/context';
 	import { BaseBranchService } from '$lib/vbranches/baseBranch';
